requirements.txt
flask
google-generativeai
pydantic>=2
numpy
